refactor: as snap daemon
