--- prometheus/ckp/__init__.py
@@ -1,1 +1,0 @@
-"""Charmed-Kafka custom Prometheus metrics module."""
