+++ prometheus/src/ckp/__init__.py
@@ -5,16 +5,16 @@ import time
 
 from prometheus_client import start_http_server
 
-from ckp.cli import run_bin_command
-from ckp.core import Config, Metrics
-from ckp.utils import parse_consumer_groups_output, validate_config
+from .cli import run_bin_command
+from .core import Config, Metrics
+from .utils import parse_consumer_groups_output, validate_config
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger("charmed-kafka-prom")
 
 
-def loop(config: Config):
-    """Main metric extraction loop."""
+def iterate(config: Config):
+    """Main metric extraction iteration."""
     raw = run_bin_command(
         "consumer-groups",
         [
@@ -33,22 +33,20 @@ def loop(config: Config):
         Metrics.CONSUMER_LAG.labels(item.group, item.topic, item.partition).set(item.lag)
 
 
-if __name__ == "__main__":
-    # if not snap.ensure(Constants.SNAP, snap.SnapState.Present.value):
-    #     print(f"{SNAP} should be present.")
-    #     sys.exit(8)
-
+def main():
+    """Main entrypoint."""
     config = validate_config()
     start_http_server(config.PORT)
     while 1:
         t0 = time.perf_counter()
         try:
-            loop(config=config)
+            iterate(config=config)
         except Exception as e:
-            logger.error(f"Metric extraction failed: {e}")
+            extended_msg = f"{getattr(e, 'stdout', '')} {getattr(e, 'stderr', '')}"
+            logger.error(f"Metric extraction failed: {e} {extended_msg}")
         finally:
             took = time.perf_counter() - t0
-            logger.info(f"Metrics extraction loop finished in {took} seconds")
+            logger.info(f"Metrics extraction finished in {took} seconds")
 
         _sleep = max(config.CYCLE - took, 0.1)
         time.sleep(_sleep)
