+++ prometheus/tox.ini
@@ -2,8 +2,8 @@
 env_list = lint, main
 
 [vars]
-src_path = {tox_root}/ckp
-all_path = main.py {[vars]src_path} 
+src_path = {tox_root}/src
+all_path = {[vars]src_path} 
 
 [testenv]
 set_env =
